update the location from client

+++ api/DTOs/UpdateLocationRequestDTO.cs
@@ -9,8 +9,8 @@ namespace api.DTOs
         public string Title { get; set; } = string.Empty;
         public string StreetAddress { get; set; } = string.Empty;
         public string Username { get; set; } = string.Empty;
-        public int Radius { get; set; }
-        public string RadiusUnit { get; set; } = string.Empty;
+        public float Radius { get; set; }
+        public string RadiusUnit { get; set; } = string.Empty; 
         public List<Reminder> Reminders { get; set; } = new List<Reminder>();
     }
 }
